Refactor DeepSeek V4 reasoning helper imports
